% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_beta_from_methylkit.R
\name{build_beta_from_methylkit}
\alias{build_beta_from_methylkit}
\title{build_beta_from_methylkit}
\usage{
build_beta_from_methylkit(test)
}
\arguments{
\item{test}{the result of methylkit with slopes including .data and  sample.ids}
}
\value{
data.frame
}
\description{
build_beta_from_methylkit
}
