% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flit_img_from_directory.R
\name{flit_img_from_directory}
\alias{flit_img_from_directory}
\title{flit_img_from_directory}
\usage{
flit_img_from_directory(working_dir, img_format = "png")
}
\arguments{
\item{working_dir}{the dir stores the images}

\item{img_format}{default as png}
}
\value{
null
}
\description{
flit_img_from_directory
}
