% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flop_img_from_directory.R
\name{flop_img_from_directory}
\alias{flop_img_from_directory}
\title{flop_img_from_directory}
\usage{
flop_img_from_directory(working_dir, img_format = "png")
}
\arguments{
\item{working_dir}{the dir stores the images}

\item{img_format}{default as png}
}
\value{
null
}
\description{
flop_img_from_directory
}
