% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet_color_scheme.R
\name{monet_color_scheme}
\alias{monet_color_scheme}
\title{monet_color_scheme}
\usage{
monet_color_scheme(num_colors = 5, randomgen = FALSE, seed = NULL)
}
\arguments{
\item{num_colors}{num the number of colors}

\item{randomgen}{logit generate color scheme randomly when true}

\item{seed}{num the set.seed() to keep reproducible sampling}
}
\value{
vector
}
\description{
gen color scheme from monet's works or similar to material you
}
\examples{
monet_color_scheme(5)
}
