% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_img_from_directory.R
\name{resize_img_from_directory}
\alias{resize_img_from_directory}
\title{resize_img_from_directory}
\usage{
resize_img_from_directory(
  working_dir,
  img_format = "png",
  resize_w = 224,
  resize_h = 224
)
}
\arguments{
\item{working_dir}{the dir stores the images}

\item{img_format}{default as png}

\item{resize_w}{the size}

\item{resize_h}{the size}
}
\value{
null
}
\description{
resize_img_from_directory
}
