% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_img_from_directory.R
\name{rotate_img_from_directory}
\alias{rotate_img_from_directory}
\title{rotate_img_from_directory}
\usage{
rotate_img_from_directory(working_dir, x, img_format = "png")
}
\arguments{
\item{working_dir}{the dir stores the images}

\item{x}{the rotated angle}

\item{img_format}{default as png}
}
\value{
null
}
\description{
rotate_img_from_directory
}
